requirements.txt
# discord.py with voice support (PyNaCl provides Opus encryption)
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0
